Added generic interface for AggregationResultsFactory
